requirements.txt
# Core Django
Django~=5.2.2
asgiref~=3.8.1
sqlparse~=0.5.1
python-decouple~=3.8

# Development Tools
django-browser-reload~=1.13.0
django-debug-toolbar~=4.2.0
django-extensions~=3.2.3

# API & REST Framework
djangorestframework~=3.14.0
djangorestframework-simplejwt~=5.3.0
drf-spectacular~=0.26.5
django-cors-headers~=4.3.1

# Database & ORM
psycopg2-binary~=2.9.7
django-redis~=5.4.0
celery~=5.3.4
redis~=5.0.1

# AI Services
openai~=1.3.7
anthropic~=0.7.7
google-generativeai~=0.3.2
requests~=2.31.0
httpx~=0.25.2

# Authentication & Security
django-oauth-toolkit~=1.7.1
django-allauth~=0.57.0
PyJWT~=2.8.0
cryptography~=41.0.7

# Payment Processing
stripe~=7.8.0
paypalrestsdk~=1.13.3

# WebSockets & Real-time
channels~=4.0.0
channels-redis~=4.1.0
websockets~=12.0

# File Processing & Storage
Pillow~=10.1.0
django-storages~=1.14.2
boto3~=1.34.0
cloudinary~=1.36.0

# Monitoring & Analytics
sentry-sdk~=1.38.0
django-prometheus~=2.3.1
newrelic~=9.2.0

# Task Queue
celery[redis]~=5.3.4
django-celery-beat~=2.5.0
django-celery-results~=2.5.1

# Utilities
python-slugify~=8.0.1
faker~=20.1.0
python-dateutil~=2.8.2
pytz~=2023.3
whitenoise~=6.6.0

# Production Server
gunicorn~=21.2.0
uvicorn[standard]~=0.24.0

# Data Processing
pandas~=2.1.4
numpy~=1.25.2
matplotlib~=3.8.2

# Email
django-anymail~=10.2
